requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
python-decouple==3.8
Pillow==11.2.1
django-tailwind==4.0.1
drf-yasg==1.21.7
django-filter==24.1
djangorestframework-filters==1.0.0.dev2
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.34
django-redis==5.4.0
redis==5.0.1
django-celery==3.4.0
celery==5.3.6
sentry-sdk==1.39.1
python-magic-bin==0.4.14; sys_platform == 'win32'
psutil==5.9.8
django-health-check==3.17.0
django-prometheus==2.3.1
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-cleanup==8.0.0
django-ratelimit==4.0.0
django-axes==6.3.0
django-admin-honeypot==1.1.0
django-defender==0.8.4
django-ipware==6.0.0
django-user-agents==0.4.0
django-request-logging==0.7.1
django-auditlog==2.3.0
django-simple-history==3.4.0
django-admin-interface==0.25.0
django-admin-rangefilter==0.19.0
django-admin-list-filter-dropdown==1.0.1
django-admin-autocomplete-filter==0.7.0
django-admin-tools==0.9.3
django-admin-honeypot==1.1.0
django-admin-interface==0.25.0
django-admin-rangefilter==0.19.0
django-admin-list-filter-dropdown==1.0.1
django-admin-autocomplete-filter==0.7.0
django-admin-tools==0.9.3 